b2p

+++ BTCPayServer.Plugins.B2PCentral/Services/B2PCentralPluginService.cs
@@ -391,7 +391,7 @@ public class B2PCentralPluginService(B2PCentralPluginDbContextFactory pluginDbCo
         try
         {
             await using var btcPayCtx = btcPayDbContextFactory.CreateContext();
-            return !await btcPayCtx.Payouts.AnyAsync(a => a.State <= PayoutState.InProgress && a.PayoutMethodId == "BTC-CHAIN");
+            return !await btcPayCtx.Payouts.AnyAsync(a => (int)a.State <= (int)PayoutState.InProgress && a.PayoutMethodId == "BTC-CHAIN");
         }
         catch (Exception e)
         {
